backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==4.2.28
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
exceptiongroup==1.3.1
flower==2.0.1
humanize==4.13.0
idna==3.11
kombu==5.6.2
packaging==26.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
